Zaheer release 1

+++ src/app/pages/users/users.component.ts
@@ -1,25 +1,29 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatSort } from '@angular/material/sort';
-import { MatTableDataSource } from '@angular/material/table';
-import { Router } from '@angular/router';
-import { User } from 'src/app/models/user';
-import { TokenStorageService } from 'src/app/_services/token-storage.service';
-import { AuthService } from './../../_services/auth.service';
+import {Component, OnInit, ViewChild} from '@angular/core';
+import {MatPaginator} from '@angular/material/paginator';
+import {MatSort} from '@angular/material/sort';
+import {MatTableDataSource} from '@angular/material/table';
+import {Router} from '@angular/router';
+import {User} from 'src/app/models/user';
+import {TokenStorageService} from 'src/app/_services/token-storage.service';
+import {AuthService} from '../../_services/auth.service';
+import Swal from "sweetalert2";
+import {MatDialog} from "@angular/material/dialog";
+import {AdduserformComponent} from "../../modal/adduserform/adduserform.component";
+
 @Component({
   selector: 'app-users',
   templateUrl: './users.component.html',
   styleUrls: ['./users.component.css']
 })
 export class UsersComponent implements OnInit {
-  columnsToDisplay = ["id", "username", "fullname", "email", "status", "roles", "action"];
+  columnsToDisplay = ["id", "username", "name", "email", "status", "roles", "action"];
   users: User[] = null;
   dataSource: MatTableDataSource<User> = null;
-
   @ViewChild(MatSort) sort: MatSort;
   @ViewChild(MatPaginator) paginator: MatPaginator;
 
-  constructor(private token: TokenStorageService, private authService: AuthService, private router: Router) { }
+  constructor(private token: TokenStorageService, private authService: AuthService, private router: Router,public dialog: MatDialog) {
+  }
 
   ngOnInit() {
     if (!this.token.isAdmin()) {
@@ -27,6 +31,7 @@ export class UsersComponent implements OnInit {
     }
     this.getAllUsers()
   }
+
   getAllUsers() {
     this.authService.getAllUsers().subscribe(
       data => {
@@ -42,6 +47,7 @@ export class UsersComponent implements OnInit {
       }
     );
   }
+
   public doFilter = (value: string, type: String) => {
     switch (type) {
       case 'user':
@@ -49,15 +55,58 @@ export class UsersComponent implements OnInit {
 
     }
   }
+
   deleteUser(user: User): any {
-    this.authService.deleteUser(user).subscribe(
-      data => {
-        this.getAllUsers()
-      },
-      err => {
-        console.log(err);
+
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.authService.deleteUser(user).subscribe(
+          data => {
+            this.getAllUsers();
+            Swal.fire(
+              'Deleted!',
+              'Account has been deleted.',
+              'success'
+            )
+          },
+          err => {
+            console.log(err);
+          }
+        );
       }
-    );
+    })
+
+  }
+
+
+  openDialog(user?: User): void {
+    const dialogRef = this.dialog.open(AdduserformComponent, {
+      width: '530px',
+    });
+    dialogRef.afterClosed().subscribe(res => {
+      console.log("user registered!")
+      // if (JSON.stringify(user) != JSON.stringify(res)) {
+      //   if (user.id != res.id)
+      //     console.log("error")
+      //   else {
+          this.getAllUsers();
+          // this.updateProduct(res)
+          // this.refreshproduct()
+        // }
+      // }
+
+
+
+    });
   }
 
 }
